PP-5146 - Extract this function so we can use it in the controller later
